FileAttachment model now works proper
FileAttachment model now returns URL when value is formated. Further, while FileAttachment saves the file, it keeps an extra field valueid to track the id. For the development and test environments, the upload directory is test/fixtures/files. New instance and its detail vlaues are added for Persons entity (101) of database 6 of account 1. The uploaded picture is placed at test/fixtures/files/1/6/101/202/1215. The whole functionality is checked the provided link returns the uploaded file. Though uploads from the form aare yet to be checked.