FileAttachment upload works now
FileAttachmentController is a bit more robust in case it cannot find the uploaded file (mostly due to change in the location of storage directory)

Entity model method has_file_attachment_details? is modified to incorporate the madb_file_attachment type

FileAttachment Model now properly updates the attributes in save_file which it was not doing before. In save_file, update_attribute is called to save the updated values which in turn invokes value=() which is actually modfied to prevent recusion.